graphql/requirements.txt
anyio==4.0.0
ariadne==0.20.1
blinker==1.6.3
bcrypt==4.0.1
click==8.1.7
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Session==0.5.0
Flask-SQLAlchemy==3.1.1
google-cloud-storage==2.12.0
graphql-core==3.2.3
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.0
pydantic_settings==2.0.3
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.31.1
typing_extensions==4.8.0
Werkzeug==3.0.0
